+++ fla/ops/generalized_delta_rule/dplr/fused_recurrent.py
@@ -227,7 +227,7 @@ def fused_recurrent_dplr_delta_rule(
             b of shape `[B, T, H, K]`.
         gk (torch.Tensor):
             gk of shape `[B, T, H, K]`. decay term in log space!
-        scale (Optional[int]):
+        scale (Optional[float]):
             Scale factor for the RetNet attention scores.
             If not provided, it will default to `1 / sqrt(K)`. Default: `None`.
         initial_state (Optional[torch.Tensor]):
